% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_random.R
\name{send_random}
\alias{send_random}
\title{Title}
\usage{
send_random(
  subject = "This is a test.",
  month = NULL,
  body = "Surprise!",
  email = options()$email
)
}
\arguments{
\item{subject}{character string indicating subject line of email}

\item{month}{character string indicating month to send email}

\item{body}{character string indicating body of email}

\item{email}{from email address, defaults to options()$email}
}
\description{
Title
}
\examples{

gmailr::gm_auth_configure()
send_random("feb")

}
